added some more test and made installable

+++ luckysocial.py
@@ -1,34 +1,56 @@
 #!/usr/bin/env python3
 
+import os
 import re
 import csv
 import sys
 import time
 import random
+import argparse
 import collections
 import requests_html
 
-ua = "Mozilla/5.0 (Linux; Android 7.0; SM-G930V Build/NRD90M) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/59.0.3071.125 Mobile Safari/537.36"
-
-http = requests_html.HTMLSession()
-new_fields = ["homepage", "twitter", "facebook", "instagram", "youtube", "rss"]
-
 def main():
-    csv_file = sys.argv[1]
-    in_csv = csv.DictReader(open(csv_file))
-    fieldnames = in_csv.fieldnames + new_fields
+    parser = argparse.ArgumentParser(description="look up social media accounts")
+    parser.add_argument('input', help='Either a name to lookup or a CSV file')
+    parser.add_argument('--name-col', default='name', help='The column containing the name')
+    parser.add_argument('--url-col', default=None, help='A column containing the URL to lookup')
+    args = parser.parse_args()
 
-    new_csv_file = csv_file.replace('.csv', '-new.csv')
+    # if a name was passed in as input print what we know about it and exit
+
+    if not os.path.isfile(args.input):
+        info = {"name": args.input}
+        info.update(lookup(args.input))
+        if info:
+            print_info(info)
+        print()
+        sys.exit()
+
+    # otherwise we are processing a CSV
+
+    in_csv = csv.DictReader(open(args.input))
+    if not args.url_col and args.name_col not in in_csv.fieldnames:
+        parser.error('{} column not found in {}'.format(args.name_col, args.input))
+    if args.url_col and args.url_col not in in_csv.fieldnames:
+        parser.error('{} column not found in {}'.format(args.url_col, args.input))
+
+    fieldnames = in_csv.fieldnames + new_fields
+    new_csv_file = args.input.replace('.csv', '-new.csv')
     out_csv = csv.DictWriter(open(new_csv_file, "w"), fieldnames=fieldnames)
     out_csv.writeheader()
 
     for row in in_csv:
-        name = row["name"]
-        info = lookup(name)
+        if not args.url_col:
+            info = lookup(row[args.name_col])
+            # TODO: make the sleep optional?
+            # it is hear to prevent getting blocked by Google for scraping
+            time.sleep(random.randint(0, 3))
+        else:
+            info = get_social(row[args.url_col])
         row.update(info)
         out_csv.writerow(row)
-        print(row)
-        time.sleep(random.randint(0, 3))
+        print_info(row)
 
 def lookup(name):
     result = {"homepage": get_homepage(name)}
@@ -50,13 +72,16 @@ def get_social(url):
     if not url:
         return no_result
 
+    if not url.startswith('http'):
+        url = 'http://' + url
+
     try:
         doc = http.get(url)
     except Exception as e:
         return no_result
 
     return {
-        "twitter": get_meta(doc, 'twitter:creator') or find_url(doc, r".*twitter.com/(?:#!/)?([a-z0-9_]+).*", "https://twitter.com/", ["intent"]),
+        "twitter": get_meta(doc, 'twitter:creator') or find_url(doc, r".*twitter.com/(?:#!/)?([a-z0-9_]+)/?$", "https://twitter.com/", ["intent"]),
         "facebook": find_url(doc, r".*facebook.com/([a-z0-9_]+)/?$", "https://www.facebook.com/"),
         "instagram": find_url(doc, r".*instagram.com/([a-z0-9_]+)/?$", "https://www.instagram.com/"),
         "youtube": find_url(doc, r".*youtube.com/user/([a-z0-9_]+)/?$", "https://www.youtube.com/user/"),
@@ -65,9 +90,9 @@ def get_social(url):
 
 def get_meta(doc, name):
     meta = doc.html.find('meta[name="{}"]'.format(name), first=True)
-    if meta:
-        # TODO: create url from bare screen name
-        return meta.attrs['content']
+    if meta and meta.attrs['content']:
+        # TODO: make not twitter specific
+        return meta.attrs['content'].replace('@', 'https://twitter.com/')
     else:
         return None
  
@@ -88,6 +113,17 @@ def get_rss(doc):
         if 'href' in link.attrs and 'comments' not in link.attrs['href']:
             return link.attrs['href']
     return None
+
+def print_info(info):
+    print()
+    for k, v in info.items():
+        if v:
+            print('{}: {}'.format(k, v))
+
+ua = "Mozilla/5.0 (Linux; Android 7.0; SM-G930V Build/NRD90M) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/59.0.3071.125 Mobile Safari/537.36"
+
+http = requests_html.HTMLSession()
+new_fields = ["homepage", "twitter", "facebook", "instagram", "youtube", "rss"]
 
 if __name__ == "__main__":
     main()
